Fix wrong year options in edit form
- Add YearDropdown component and use the same logic in AddItemForm and in edit form (Inventory component)
